Move VNet to new home
