+++ web/packages/teleterm/src/ui/TopBar/Vnet/index.ts
@@ -1,6 +1,6 @@
 /**
  * Teleport
- * Copyright (C) 2024 Gravitational, Inc.
+ * Copyright (C) 2025 Gravitational, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -16,6 +16,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from './VnetPanel';
-export * from './vnetContext';
-export * from './useVnetLauncher';
+export { Vnet } from './Vnet';
