--- web/packages/teleterm/src/ui/Vnet/VnetSliderStep.story.tsx
@@ -1,481 +1,0 @@
-/**
- * Teleport
- * Copyright (C) 2024 Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-import { Meta, StoryObj } from '@storybook/react-vite';
-import { useEffect } from 'react';
-
-import { Box } from 'design';
-import { Timestamp } from 'gen-proto-ts/google/protobuf/timestamp_pb';
-import {
-  ConnectionStat,
-  RecentConnection,
-  RecentConnectionKind,
-  WindowsServiceStatus,
-} from 'gen-proto-ts/teleport/lib/teleterm/vnet/v1/vnet_service_pb';
-import {
-  CheckAttemptStatus,
-  CheckReportStatus,
-} from 'gen-proto-ts/teleport/lib/vnet/diag/v1/diag_pb';
-import { usePromiseRejectedOnUnmount } from 'shared/utils/wait';
-
-import { MockedUnaryCall } from 'teleterm/services/tshd/cloneableClient';
-import { makeRootCluster } from 'teleterm/services/tshd/testHelpers';
-import { makeReport } from 'teleterm/services/vnet/testHelpers';
-import { MockAppContextProvider } from 'teleterm/ui/fixtures/MockAppContextProvider';
-import { MockAppContext } from 'teleterm/ui/fixtures/mocks';
-import { ConnectionsContextProvider } from 'teleterm/ui/TopBar/Connections/connectionsContext';
-
-import { useVnetContext, VnetContextProvider } from './vnetContext';
-import { VnetSliderStep as Component } from './VnetSliderStep';
-
-type StoryProps = {
-  startVnet: 'success' | 'error' | 'processing';
-  autoStart: boolean;
-  appDnsZones: string[];
-  clusters: string[];
-  recentConnections: RecentConnection[];
-  connectionStats: ConnectionStat[];
-  sshConfigured: boolean;
-  fetchStatus:
-    | 'success'
-    | 'error'
-    | 'processing'
-    | 'processing-with-previous-results';
-  runDiagnostics: 'success' | 'error' | 'processing';
-  diagReport: 'ok' | 'issues-found' | 'failed-checks';
-  isWorkspacePresent: boolean;
-  unexpectedShutdown: boolean;
-  installTimeRequirementsCheck:
-    | 'success'
-    | 'windows-service-not-installed'
-    | 'windows-service-version-mismatch';
-};
-
-const defaultArgs: StoryProps = {
-  startVnet: 'success',
-  autoStart: true,
-  appDnsZones: ['teleport.example.com', 'company.test'],
-  clusters: ['teleport.example.com'],
-  recentConnections: [],
-  connectionStats: [],
-  sshConfigured: false,
-  fetchStatus: 'success',
-  runDiagnostics: 'success',
-  diagReport: 'ok',
-  isWorkspacePresent: true,
-  unexpectedShutdown: false,
-  installTimeRequirementsCheck: 'success',
-};
-
-const meta: Meta<StoryProps> = {
-  title: 'Teleterm/Vnet/VnetSliderStep',
-  component: VnetSliderStep,
-  decorators: [
-    Story => {
-      return (
-        <Box width={396} bg="levels.elevated">
-          <Story />
-        </Box>
-      );
-    },
-  ],
-  argTypes: {
-    startVnet: {
-      control: { type: 'inline-radio' },
-      options: ['success', 'error', 'processing'],
-    },
-    appDnsZones: {
-      control: { type: 'object' },
-    },
-    clusters: {
-      control: { type: 'object' },
-    },
-    fetchStatus: {
-      control: { type: 'inline-radio' },
-      options: [
-        'success',
-        'error',
-        'processing',
-        'processing-with-previous-results',
-      ],
-    },
-    runDiagnostics: {
-      control: { type: 'inline-radio' },
-      options: ['success', 'error', 'processing'],
-    },
-    diagReport: {
-      control: { type: 'inline-radio' },
-      options: ['ok', 'issues-found', 'failed-checks'],
-    },
-    isWorkspacePresent: {
-      description:
-        "If there's no workspace, the button to open the diag report is disabled.",
-    },
-    installTimeRequirementsCheck: {
-      control: { type: 'radio' },
-      options: [
-        'success',
-        'windows-service-not-installed',
-        'windows-service-version-mismatch',
-      ],
-      description: 'VNet-related checks performed before startup.',
-    },
-  },
-  render: props => <VnetSliderStep {...props} />,
-};
-export default meta;
-
-function VnetSliderStep(props: StoryProps) {
-  const appContext = new MockAppContext();
-
-  if (props.installTimeRequirementsCheck === 'windows-service-not-installed') {
-    appContext.vnet.checkInstallTimeRequirements = () =>
-      new MockedUnaryCall({
-        status: {
-          oneofKind: 'windowsServiceStatus' as const,
-          windowsServiceStatus: WindowsServiceStatus.DOES_NOT_EXIST,
-        },
-      });
-  }
-  if (
-    props.installTimeRequirementsCheck === 'windows-service-version-mismatch'
-  ) {
-    appContext.vnet.checkInstallTimeRequirements = () =>
-      new MockedUnaryCall({
-        status: {
-          oneofKind: 'windowsServiceStatus' as const,
-          windowsServiceStatus: WindowsServiceStatus.VERSION_MISMATCH,
-        },
-      });
-  }
-
-  if (props.isWorkspacePresent) {
-    appContext.addRootCluster(makeRootCluster());
-  }
-
-  if (props.autoStart) {
-    appContext.statePersistenceService.putState({
-      ...appContext.statePersistenceService.getState(),
-      vnet: { autoStart: true, hasEverStarted: true },
-    });
-    appContext.workspacesService.setState(draft => {
-      draft.isInitialized = true;
-    });
-  }
-
-  const pendingPromise = usePromiseRejectedOnUnmount();
-
-  if (props.startVnet === 'processing') {
-    appContext.vnet.start = () => pendingPromise;
-  } else {
-    appContext.vnet.start = () => {
-      if (props.startVnet === 'success' && props.unexpectedShutdown) {
-        setTimeout(() => {
-          appContext.unexpectedVnetShutdownListener({
-            error: 'lorem ipsum dolor sit amet',
-          });
-        }, 5);
-      }
-      return new MockedUnaryCall(
-        {},
-        props.startVnet === 'error'
-          ? new Error('something went wrong')
-          : undefined
-      );
-    };
-  }
-
-  if (props.fetchStatus === 'processing') {
-    appContext.vnet.getServiceInfo = () => pendingPromise;
-  } else {
-    let firstCall = true;
-    appContext.vnet.getServiceInfo = () => {
-      if (props.fetchStatus === 'processing-with-previous-results') {
-        if (firstCall) {
-          firstCall = false;
-          return new MockedUnaryCall({
-            appDnsZones: props.appDnsZones,
-            clusters: props.clusters,
-            sshConfigured: props.sshConfigured,
-            vnetSshConfigPath:
-              '/Users/user/Library/Application Support/Teleport Connect/tsh/vnet_ssh_config',
-          });
-        }
-        return pendingPromise;
-      }
-
-      return new MockedUnaryCall(
-        {
-          appDnsZones: props.appDnsZones,
-          clusters: props.clusters,
-          sshConfigured: props.sshConfigured,
-          vnetSshConfigPath:
-            '/Users/user/Library/Application Support/Teleport Connect/tsh/vnet_ssh_config',
-        },
-        props.fetchStatus === 'error'
-          ? new Error('something went wrong')
-          : undefined
-      );
-    };
-  }
-
-  appContext.vnet.getRecentConnections = (() => ({
-    responses: {
-      onMessage: (
-        callback: (response: { connections: RecentConnection[] }) => void
-      ) => {
-        callback({ connections: props.recentConnections });
-        return () => {};
-      },
-      onNext: () => () => {},
-      onComplete: () => () => {},
-      onError: () => () => {},
-    },
-    then: () => Promise.resolve(),
-  })) as unknown as typeof appContext.vnet.getRecentConnections;
-
-  appContext.vnet.getConnectionStats = (() => ({
-    responses: {
-      onMessage: (
-        callback: (response: { stats: ConnectionStat[] }) => void
-      ) => {
-        callback({ stats: props.connectionStats });
-        return () => {};
-      },
-      onNext: () => () => {},
-      onComplete: () => () => {},
-      onError: () => () => {},
-    },
-    then: () => Promise.resolve(),
-  })) as unknown as typeof appContext.vnet.getConnectionStats;
-
-  if (props.runDiagnostics === 'processing') {
-    appContext.vnet.runDiagnostics = () => pendingPromise;
-  } else {
-    const report = makeReport();
-    const checkAttempt = report.checks[0];
-
-    if (props.diagReport === 'issues-found') {
-      checkAttempt.checkReport.status = CheckReportStatus.ISSUES_FOUND;
-    }
-
-    if (props.diagReport === 'failed-checks') {
-      checkAttempt.status = CheckAttemptStatus.ERROR;
-      checkAttempt.error = 'something went wrong';
-      checkAttempt.checkReport = undefined;
-    }
-
-    appContext.vnet.runDiagnostics = () =>
-      new MockedUnaryCall(
-        { report },
-        props.runDiagnostics === 'error'
-          ? new Error('something went wrong')
-          : undefined
-      );
-  }
-
-  return (
-    <MockAppContextProvider
-      appContext={appContext}
-      // Completely re-mount everything when controls change. This ensures that effects to fetch
-      // data are fired again.
-      key={JSON.stringify(props)}
-    >
-      <ConnectionsContextProvider>
-        <VnetContextProvider>
-          {props.fetchStatus === 'processing-with-previous-results' && (
-            <RerequestServiceInfo />
-          )}
-          <Component
-            refCallback={noop}
-            next={noop}
-            prev={noop}
-            hasTransitionEnded
-            stepIndex={1}
-            flowLength={2}
-          />
-        </VnetContextProvider>
-      </ConnectionsContextProvider>
-    </MockAppContextProvider>
-  );
-}
-
-const RerequestServiceInfo = () => {
-  const { serviceInfoAttempt, refreshServiceInfoAttempt } = useVnetContext();
-
-  useEffect(() => {
-    if (serviceInfoAttempt.status === 'success') {
-      refreshServiceInfoAttempt();
-    }
-  }, [serviceInfoAttempt, refreshServiceInfoAttempt]);
-
-  return null;
-};
-
-const noop = () => {};
-
-export const CloudCustomer: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: ['example.teleport.sh'],
-    clusters: ['example.teleport.sh'],
-  },
-};
-
-export const SelfHostedWithDifferentClusterName: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: ['teleport.example.com'],
-    clusters: ['teleport-example'],
-  },
-};
-
-export const SelfHostedWithEqualNameAndLeaf: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: ['teleport.example.com', 'leaf.example.com'],
-    clusters: ['teleport.example.com', 'leaf.example.com'],
-  },
-};
-
-export const SelfHostedWithEqualNameAndDifferentLeaf: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: ['teleport.example.com', 'leaf.example.com'],
-    clusters: ['teleport.example.com', 'teleport-leaf'],
-  },
-};
-
-export const SelfHostedWithEqualNameAndCustomDNSZones: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: ['teleport.example.com', 'company.com', 'apps.company'],
-    clusters: ['teleport.example.com'],
-  },
-};
-
-export const WithRecentConnections: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    recentConnections: [
-      {
-        kind: RecentConnectionKind.APP,
-        cluster: 'teleport.example.com',
-        leafCluster: '',
-        displayName: 'grafana.teleport.example.com',
-        lastConnected: Timestamp.fromDate(new Date(Date.now() - 15 * 1000)),
-        lastClientProcessPath: '/usr/bin/curl',
-      },
-      {
-        kind: RecentConnectionKind.SSH,
-        cluster: 'teleport.example.com',
-        leafCluster: '',
-        displayName: 'node-01.teleport.example.com',
-        lastConnected: Timestamp.fromDate(new Date(Date.now() - 5 * 60 * 1000)),
-        lastClientProcessPath: '/Applications/iTerm.app/Contents/MacOS/iTerm2',
-      },
-      {
-        kind: RecentConnectionKind.DATABASE,
-        cluster: 'teleport.example.com',
-        leafCluster: 'leaf.example.com',
-        displayName: 'postgres-main.leaf.example.com',
-        lastConnected: Timestamp.fromDate(
-          new Date(Date.now() - 42 * 60 * 1000)
-        ),
-        // No process identified (e.g. non-macOS), so no icon or "Opened by".
-        lastClientProcessPath: '',
-      },
-    ],
-  },
-};
-
-export const WithConnectionStats: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    connectionStats: [
-      {
-        kind: RecentConnectionKind.APP,
-        cluster: 'teleport.example.com',
-        leafCluster: '',
-        displayName: 'grafana.teleport.example.com',
-        port: 0,
-        successfulConnections: 12n,
-        failedConnections: 0n,
-        bytesTx: 1_300_000n,
-        bytesRx: 15_700_000n,
-        bytesTxPerSec: 12_000n,
-        bytesRxPerSec: 250_000n,
-      },
-      {
-        kind: RecentConnectionKind.APP,
-        cluster: 'teleport.example.com',
-        leafCluster: '',
-        displayName: 'multiport.teleport.example.com',
-        port: 8443,
-        successfulConnections: 2n,
-        failedConnections: 3n,
-        bytesTx: 4_200n,
-        bytesRx: 0n,
-        bytesTxPerSec: 0n,
-        bytesRxPerSec: 0n,
-      },
-      {
-        kind: RecentConnectionKind.SSH,
-        cluster: 'teleport.example.com',
-        leafCluster: '',
-        displayName: 'node-01.teleport.example.com:0',
-        port: 0,
-        successfulConnections: 4n,
-        failedConnections: 1n,
-        bytesTx: 52_000n,
-        bytesRx: 1_100_000n,
-        bytesTxPerSec: 0n,
-        bytesRxPerSec: 0n,
-      },
-      {
-        kind: RecentConnectionKind.DATABASE,
-        cluster: 'teleport.example.com',
-        leafCluster: 'leaf.example.com',
-        displayName: 'postgres-main.leaf.example.com',
-        port: 0,
-        successfulConnections: 7n,
-        failedConnections: 0n,
-        bytesTx: 830_000n,
-        bytesRx: 9_400_000n,
-        bytesTxPerSec: 1_500n,
-        bytesRxPerSec: 48_000n,
-      },
-    ],
-  },
-};
-
-export const SelfHostedWithManyLeavesAndZones: StoryObj<StoryProps> = {
-  args: {
-    ...defaultArgs,
-    appDnsZones: [
-      'teleport.example.com',
-      'leaf.example.com',
-      'second-leaf.example.com',
-      'company.com',
-    ],
-    clusters: [
-      'teleport.example.com',
-      'teleport-leaf',
-      'second-leaf.example.com',
-    ],
-  },
-};
